refactor(sync): extract hlc_guard out of apply into apply/hlc_guard.rs
